+++ sdk/communication/Azure.Communication.Identity/src/CommunicationIdentityClient.cs
@@ -281,6 +281,45 @@ namespace Azure.Communication.Identity
             }
         }
 
+        /// <summary>Exchange a teams token for a new ACS access token.</summary>
+        /// <param name="teamsToken">Teams token to acquire a new ACS access token.</param>
+        /// <param name="cancellationToken">The cancellation token to use.</param>
+        /// <exception cref="RequestFailedException">The server returned an error.</exception>
+        public virtual Response<AccessToken> ExchangeTeamsToken(string teamsToken, CancellationToken cancellationToken = default)
+        {
+            using DiagnosticScope scope = _clientDiagnostics.CreateScope($"{nameof(CommunicationIdentityClient)}.{nameof(ExchangeTeamsToken)}");
+            scope.Start();
+            try
+            {
+                Response<CommunicationIdentityAccessToken> response = RestClient.ExchangeTeamsUserAccessToken(teamsToken, cancellationToken);
+                return Response.FromValue(new AccessToken(response.Value.Token, response.Value.ExpiresOn), response.GetRawResponse());
+            }
+            catch (Exception ex)
+            {
+                scope.Failed(ex);
+                throw;
+            }
+        }
+
+        /// <summary>Asynchronously exchange a teams token for a new ACS access token.</summary>
+        /// <param name="teamsToken">Teams token to acquire a new ACS access token.</param>
+        /// <param name="cancellationToken">The cancellation token to use.</param>
+        public virtual async Task<Response<AccessToken>> ExchangeTeamsTokenAsync(String teamsToken, CancellationToken cancellationToken = default)
+        {
+            using DiagnosticScope scope = _clientDiagnostics.CreateScope($"{nameof(CommunicationIdentityClient)}.{nameof(ExchangeTeamsToken)}");
+            scope.Start();
+            try
+            {
+                Response<CommunicationIdentityAccessToken> response = await RestClient.ExchangeTeamsUserAccessTokenAsync(teamsToken, cancellationToken).ConfigureAwait(false);
+                return Response.FromValue(new AccessToken(response.Value.Token, response.Value.ExpiresOn), response.GetRawResponse());
+            }
+            catch (Exception ex)
+            {
+                scope.Failed(ex);
+                throw;
+            }
+        }
+
         private static T AssertNotNull<T>(T argument, string argumentName)
             where T : class
         {