+++ sdk/metricsadvisor/Azure.AI.MetricsAdvisor/src/Models/AnomalyDetection/AnomalyIncident.cs
@@ -17,8 +17,9 @@ namespace Azure.AI.MetricsAdvisor.Models
     [CodeGenSuppress("Property")]
     public partial class AnomalyIncident
     {
-        internal AnomalyIncident(string metricId, string detectionConfigurationId, string id, DateTimeOffset startTime, DateTimeOffset endTime, SeriesIdentity rootNode, IncidentProperty property)
+        internal AnomalyIncident(Guid? dataFeedId, string metricId, string detectionConfigurationId, string id, DateTimeOffset startTime, DateTimeOffset endTime, SeriesIdentity rootNode, IncidentProperty property)
         {
+            DataFeedId = dataFeedId;
             MetricId = metricId;
             DetectionConfigurationId = detectionConfigurationId;
             Id = id;
@@ -93,5 +94,7 @@ namespace Azure.AI.MetricsAdvisor.Models
         /// enough to make a prediction, or if the anomaly was not detected by a <see cref="SmartDetectionCondition"/>.
         /// </summary>
         public double? ExpectedValueOfRootNode { get; }
+
+        internal Guid? DataFeedId { get; }
     }
 }